refactor: change packages and delete unused class
