+++ src/main/java/com/cnubot/cnubotserver/board/controller/BoardController.java
@@ -1,12 +1,11 @@
-package com.cnubot.cnubotserver.domain.controller;
+package com.cnubot.cnubotserver.board.controller;
 
 
-import com.cnubot.cnubotserver.domain.entity.Board;
-import com.cnubot.cnubotserver.domain.service.BoardService;
+import com.cnubot.cnubotserver.board.entity.Board;
+import com.cnubot.cnubotserver.board.service.BoardService;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
