+++ src/main/java/com/cnubot/cnubotserver/board/repository/BoardRepository.java
@@ -1,7 +1,7 @@
-package com.cnubot.cnubotserver.domain.repository;
+package com.cnubot.cnubotserver.board.repository;
 
-import com.cnubot.cnubotserver.domain.entity.Board;
-import com.cnubot.cnubotserver.domain.entity.DepthSecond;
+import com.cnubot.cnubotserver.board.entity.Board;
+import com.cnubot.cnubotserver.board.entity.DepthSecond;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
