+++ src/main/java/com/cnubot/cnubotserver/board/service/BoardServiceImpl.java
@@ -1,10 +1,9 @@
-package com.cnubot.cnubotserver.domain.service;
+package com.cnubot.cnubotserver.board.service;
 
-import com.cnubot.cnubotserver.domain.entity.Board;
-import com.cnubot.cnubotserver.domain.entity.DepthSecond;
-import com.cnubot.cnubotserver.domain.repository.BoardRepository;
+import com.cnubot.cnubotserver.board.entity.Board;
+import com.cnubot.cnubotserver.board.entity.DepthSecond;
+import com.cnubot.cnubotserver.board.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,7 +12,7 @@ import java.util.List;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class BoardServiceImpl implements BoardService{
+public class BoardServiceImpl implements BoardService {
 
     private final BoardRepository repository;
     private final BoardCrawling boardCrawling;
