--- src/main/java/com/cnubot/cnubotserver/domain/controller/AdminController.java
@@ -1,12 +1,0 @@
-package com.cnubot.cnubotserver.domain.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-import springfox.documentation.annotations.ApiIgnore;
-
-/**
- * 관리자 페이지 Controller입니다.
- */
-@ApiIgnore
-@RestController
-public class AdminController {
-}
