--- src/main/java/com/cnubot/cnubotserver/domain/controller/MainController.java
@@ -1,13 +1,0 @@
-package com.cnubot.cnubotserver.domain.controller;
-
-
-import org.springframework.web.bind.annotation.RestController;
-import springfox.documentation.annotations.ApiIgnore;
-
-/**
- *  버스와 충남대 홈페이지 소식 게시글 관련 Controller입니다.
- */
-@ApiIgnore
-@RestController
-public class MainController {
-}
