--- src/main/java/com/cnubot/cnubotserver/domain/entity/Member.java
@@ -1,5 +1,0 @@
-package com.cnubot.cnubotserver.domain.entity;
-
-
-public class Member {
-}
