+++ src/main/java/com/cnubot/cnubotserver/foodcourt/service/crawling/FoodCrawling.java
@@ -22,7 +22,7 @@ import org.springframework.stereotype.Service;
 @AllArgsConstructor
 public class FoodCrawling {
 
-    MenuRepository repository;
+    private final MenuRepository repository;
 
     public void process() {
         FoodCourt[] values = FoodCourt.values();
