+++ src/main/java/com/cnubot/cnubotserver/swagger/SwaggerConfiguration.java
@@ -1,4 +1,4 @@
-package com.cnubot.cnubotserver.domain.config;
+package com.cnubot.cnubotserver.swagger;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
